# frontend/openapi.json
{
  "components": {
    "schemas": {
      "AppliedFilters": {
        "description": "The filters actually applied after validation and defaults (default window: 30 days\nending at the database's latest trending date). Empty country/category lists mean all.",
        "properties": {
          "categories": {
            "description": "Applied categories; empty = all categories.",
            "items": {
              "type": "string"
            },
            "title": "Categories",
            "type": "array"
          },
          "countries": {
            "description": "Applied country codes; empty = all countries.",
            "items": {
              "type": "string"
            },
            "title": "Countries",
            "type": "array"
          },
          "days": {
            "title": "Days",
            "type": "integer"
          },
          "end_date": {
            "format": "date",
            "title": "End Date",
            "type": "string"
          },
          "start_date": {
            "format": "date",
            "title": "Start Date",
            "type": "string"
          }
        },
        "required": [
          "start_date",
          "end_date",
          "days",
          "countries",
          "categories"
        ],
        "title": "AppliedFilters",
        "type": "object"
      },
...
      "CategoryPerformance": {
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "engagement_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "VIDEO-LEVEL: (SUM(likes) + SUM(comments)) / SUM(views); null when views sum to 0.",
            "title": "Engagement Rate"
          },
          "trending_volume": {
            "description": "SNAPSHOT-LEVEL: filtered snapshots whose own category is this one.",
            "title": "Trending Volume",
            "type": "integer"
          },
          "unique_videos": {
            "description": "VIDEO-LEVEL: videos whose latest snapshot within the filtered dataset is in this category (sums to the overall unique video count).",
            "title": "Unique Videos",
            "type": "integer"
          },
          "views": {
            "description": "VIDEO-LEVEL: sum of those latest snapshots' view_count (once per video).",
            "title": "Views",
            "type": "integer"
          },
          "volume_share_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "SNAPSHOT-LEVEL: trending_volume / all filtered snapshots * 100.",
            "title": "Volume Share Pct"
          }
        },
        "required": [
          "category",
          "trending_volume",
          "volume_share_pct",
          "unique_videos",
          "views",
          "engagement_rate"
        ],
        "title": "CategoryPerformance",
        "type": "object"
      },
...
      "ChannelSort": {
        "enum": [
          "unique_videos",
          "trending_volume",
          "views"
        ],
        "title": "ChannelSort",
        "type": "string"
      },
...
      "EngagementAnalysis": {
        "description": "VIDEO-LEVEL engagement over each video's latest snapshot within the filtered dataset.\nHidden like counts are stored as 0 and counted as 0.",
        "properties": {
          "by_category": {
            "items": {
              "$ref": "#/components/schemas/CategoryEngagement"
            },
            "title": "By Category",
            "type": "array"
          },
          "distribution": {
            "$ref": "#/components/schemas/EngagementDistribution"
          },
          "histogram": {
            "description": "1-point buckets 0-10% plus '>= 10%'; all buckets returned.",
            "items": {
              "$ref": "#/components/schemas/HistogramBucket"
            },
            "title": "Histogram",
            "type": "array"
          },
          "scatter": {
            "description": "Top videos by views (ties: video_id).",
            "items": {
              "$ref": "#/components/schemas/ScatterPoint"
            },
            "title": "Scatter",
            "type": "array"
          },
          "totals": {
            "$ref": "#/components/schemas/EngagementTotals"
          }
        },
        "required": [
          "totals",
          "distribution",
          "histogram",
          "by_category",
          "scatter"
        ],
        "title": "EngagementAnalysis",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "description": "Stable machine-readable error code.",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ErrorDetail"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Same value as the X-Request-ID response header.",
            "title": "Request Id"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "FilteredResponse_EngagementAnalysis_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/EngagementAnalysis"
          },
          "filters": {
            "$ref": "#/components/schemas/AppliedFilters"
          }
        },
        "required": [
          "filters",
          "data"
        ],
        "title": "FilteredResponse[EngagementAnalysis]",
        "type": "object"
      },
...
      "FilteredResponse_list_CategoryPerformance__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/CategoryPerformance"
            },
            "title": "Data",
            "type": "array"
          },
          "filters": {
            "$ref": "#/components/schemas/AppliedFilters"
          }
        },
        "required": [
          "filters",
          "data"
        ],
        "title": "FilteredResponse[list[CategoryPerformance]]",
        "type": "object"
      },
      "FilteredResponse_list_ChannelPerformance__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ChannelPerformance"
            },
            "title": "Data",
            "type": "array"
          },
          "filters": {
            "$ref": "#/components/schemas/AppliedFilters"
          }
        },
        "required": [
          "filters",
          "data"
        ],
        "title": "FilteredResponse[list[ChannelPerformance]]",
        "type": "object"
      },
      "FilteredResponse_list_CountryPerformance__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/CountryPerformance"
            },
            "title": "Data",
            "type": "array"
          },
          "filters": {
            "$ref": "#/components/schemas/AppliedFilters"
          }
        },
        "required": [
          "filters",
          "data"
        ],
        "title": "FilteredResponse[list[CountryPerformance]]",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "\nRead-only analytics API over YouTube trending snapshots (PostgreSQL `app` schema) and a frozen ML model.\n\n**Metric semantics**\n* *Latest snapshot* = each video's latest snapshot **within the filtered dataset** (date range,\n  countries, categories), not its globally latest snapshot.\n* **Snapshot-level** metrics count trending appearances (country x video x day), e.g. trending volume.\n* **Video-level** metrics count each video once via its latest filtered snapshot, e.g. unique videos,\n  views, engagement rate.\n* **Country-level** metrics use each video's latest filtered snapshot in that country. Country\n  `views_of_trending_videos` are **global lifetime views** of videos that trended there; they are\n  **not additive** across countries.\n* View counts are lifetime totals: historical snapshots are **never summed** for views.\n* Period change: counts and views in **percent**; engagement rate in **percentage points**.\n\n**Authentication**: sign in via `/api/v1/auth/login` (or register). The session is an HttpOnly\ncookie; every `/api/v1` endpoint except register/login requires it. Analytics endpoints need\n`VIEW_ANALYTICS`, predictions `MAKE_PREDICTION`, member management `MANAGE_USERS`. The YouTube\ndataset is global and read-only; workspaces (tenants) own only their users and memberships.\nHealth probes are public.\n",
    "title": "YouTube Trending Intelligence API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/analytics/categories": {
      "get": {
        "description": "Snapshot-level volume uses each snapshot's own category; video-level metrics use the category of each video's latest snapshot. 'Latest snapshot' means each video's latest snapshot WITHIN THE FILTERED DATASET (the rows matching the date range, countries and categories), not its globally latest snapshot.",
        "operationId": "categories_api_v1_analytics_categories_get",
        "parameters": [
          {
            "description": "First trending date (inclusive, YYYY-MM-DD). Default: 29 days before end_date, not earlier than the first date in the data.",
            "in": "query",
            "name": "start_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "First trending date (inclusive, YYYY-MM-DD). Default: 29 days before end_date, not earlier than the first date in the data.",
              "title": "Start Date"
            }
          },
          {
            "description": "Last trending date (inclusive). Default: the latest trending date in the data.",
            "in": "query",
            "name": "end_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Last trending date (inclusive). Default: the latest trending date in the data.",
              "title": "End Date"
            }
          },
          {
            "description": "Country code; repeat for several (?country=IN&country=US). Omit for all countries.",
            "in": "query",
            "name": "country",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "maxLength": 64,
                    "type": "string"
                  },
                  "maxItems": 20,
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Country code; repeat for several (?country=IN&country=US). Omit for all countries.",
              "title": "Country"
            }
          },
          {
            "description": "Category (case-insensitive); repeat for several. Omit for all categories.",
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "maxLength": 64,
                    "type": "string"
                  },
                  "maxItems": 20,
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Category (case-insensitive); repeat for several. Omit for all categories.",
              "title": "Category"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilteredResponse_list_CategoryPerformance__"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not signed in or session expired."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Your role lacks the required permission."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request, filter or parameter."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server or database error."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database unavailable or busy."
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database statement timeout."
          }
        },
        "summary": "Category performance",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/v1/analytics/channels": {
      "get": {
        "description": "'Latest snapshot' means each video's latest snapshot WITHIN THE FILTERED DATASET (the rows matching the date range, countries and categories), not its globally latest snapshot.",
        "operationId": "channels_api_v1_analytics_channels_get",
        "parameters": [
          {
            "description": "Ranking metric (descending).",
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ChannelSort",
              "default": "unique_videos",
              "description": "Ranking metric (descending)."
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "First trending date (inclusive, YYYY-MM-DD). Default: 29 days before end_date, not earlier than the first date in the data.",
            "in": "query",
            "name": "start_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "First trending date (inclusive, YYYY-MM-DD). Default: 29 days before end_date, not earlier than the first date in the data.",
              "title": "Start Date"
            }
          },
          {
            "description": "Last trending date (inclusive). Default: the latest trending date in the data.",
            "in": "query",
            "name": "end_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Last trending date (inclusive). Default: the latest trending date in the data.",
              "title": "End Date"
            }
          },
          {
            "description": "Country code; repeat for several (?country=IN&country=US). Omit for all countries.",
            "in": "query",
            "name": "country",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "maxLength": 64,
                    "type": "string"
                  },
                  "maxItems": 20,
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Country code; repeat for several (?country=IN&country=US). Omit for all countries.",
              "title": "Country"
            }
          },
          {
            "description": "Category (case-insensitive); repeat for several. Omit for all categories.",
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "maxLength": 64,
                    "type": "string"
                  },
                  "maxItems": 20,
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Category (case-insensitive); repeat for several. Omit for all categories.",
              "title": "Category"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilteredResponse_list_ChannelPerformance__"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not signed in or session expired."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Your role lacks the required permission."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request, filter or parameter."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server or database error."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database unavailable or busy."
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database statement timeout."
          }
        },
        "summary": "Top channels",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/v1/analytics/countries": {
      "get": {
        "description": "Country-level metrics use each video's latest filtered snapshot in that country. views_of_trending_videos are GLOBAL lifetime views of videos that trended in the country: not views from the country and not additive across countries.",
        "operationId": "countries_api_v1_analytics_countries_get",
        "parameters": [
          {
            "description": "First trending date (inclusive, YYYY-MM-DD). Default: 29 days before end_date, not earlier than the first date in the data.",
            "in": "query",
            "name": "start_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "First trending date (inclusive, YYYY-MM-DD). Default: 29 days before end_date, not earlier than the first date in the data.",
              "title": "Start Date"
            }
          },
          {
            "description": "Last trending date (inclusive). Default: the latest trending date in the data.",
            "in": "query",
            "name": "end_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Last trending date (inclusive). Default: the latest trending date in the data.",
              "title": "End Date"
            }
          },
          {
            "description": "Country code; repeat for several (?country=IN&country=US). Omit for all countries.",
            "in": "query",
            "name": "country",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "maxLength": 64,
                    "type": "string"
                  },
                  "maxItems": 20,
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Country code; repeat for several (?country=IN&country=US). Omit for all countries.",
              "title": "Country"
            }
          },
          {
            "description": "Category (case-insensitive); repeat for several. Omit for all categories.",
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "maxLength": 64,
                    "type": "string"
                  },
                  "maxItems": 20,
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Category (case-insensitive); repeat for several. Omit for all categories.",
              "title": "Category"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilteredResponse_list_CountryPerformance__"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not signed in or session expired."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Your role lacks the required permission."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request, filter or parameter."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server or database error."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database unavailable or busy."
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database statement timeout."
          }
        },
        "summary": "Country performance",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/v1/analytics/engagement": {
      "get": {
        "description": "'Latest snapshot' means each video's latest snapshot WITHIN THE FILTERED DATASET (the rows matching the date range, countries and categories), not its globally latest snapshot.",
        "operationId": "engagement_api_v1_analytics_engagement_get",
        "parameters": [
          {
            "description": "Top videos by views in the scatter.",
            "in": "query",
            "name": "scatter_limit",
            "required": false,
            "schema": {
              "default": 200,
              "description": "Top videos by views in the scatter.",
              "maximum": 500,
              "minimum": 1,
              "title": "Scatter Limit",
              "type": "integer"
            }
          },
          {
            "description": "First trending date (inclusive, YYYY-MM-DD). Default: 29 days before end_date, not earlier than the first date in the data.",
            "in": "query",
            "name": "start_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "First trending date (inclusive, YYYY-MM-DD). Default: 29 days before end_date, not earlier than the first date in the data.",
              "title": "Start Date"
            }
          },
          {
            "description": "Last trending date (inclusive). Default: the latest trending date in the data.",
            "in": "query",
            "name": "end_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Last trending date (inclusive). Default: the latest trending date in the data.",
              "title": "End Date"
            }
          },
          {
            "description": "Country code; repeat for several (?country=IN&country=US). Omit for all countries.",
            "in": "query",
            "name": "country",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "maxLength": 64,
                    "type": "string"
                  },
                  "maxItems": 20,
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Country code; repeat for several (?country=IN&country=US). Omit for all countries.",
              "title": "Country"
            }
          },
          {
            "description": "Category (case-insensitive); repeat for several. Omit for all categories.",
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "maxLength": 64,
                    "type": "string"
                  },
                  "maxItems": 20,
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Category (case-insensitive); repeat for several. Omit for all categories.",
              "title": "Category"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilteredResponse_EngagementAnalysis_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not signed in or session expired."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Your role lacks the required permission."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request, filter or parameter."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unexpected server or database error."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database unavailable or busy."
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Database statement timeout."
          }
        },
        "summary": "Engagement totals, distribution, histogram, by-category spread and top-views scatter",
        "tags": [
          "analytics"
        ]
      }
    },
...
  }
}